backend/requirements.txt
fastapi
uvicorn[standard]
yfinance
httpx
python-dotenv
firebase-admin
apscheduler
google-cloud-firestore
